// src/app/model/miscChargeView.ts
export interface miscChargeViewModel {
    billNumber: string | null;
    principleAmount: number | null;
    customerNumber: string | null;
    customerName: string | null;
    nidNumber: string | null;
    tariff: string | null;
    billAmount: number | null;
    paid: number | null;
    dueDate: string | null;
    dcDate: string | null;
    paymentDate: string | null;
    rcDate: string | null;
    locationCode: string | null;
}